Move EndpointHeader to components folder
