+++ versioned_docs/version-2/1-click-signup/api-reference/endpoints.mdx
@@ -32,7 +32,7 @@ import DefaultCredentialRequestsHosted from '@site/versioned_docs/version-2/reus
 import DefaultCredentialRequestsNotHosted from '@site/versioned_docs/version-2/reusables/default-credential-requests-not-hosted.mdx'
 import RecencyRequirementAdmonition from '@site/versioned_docs/version-2/reusables/recency-requirement-admonition.mdx'
 import SandboxBirthDateAdmonition from '@site/versioned_docs/version-2/reusables/sandbox-birth-date-admonition.mdx'
-import EndpointHeader from '@site/versioned_docs/version-2/reusables/EndpointHeader.jsx'
+import EndpointHeader from '@site/src/components/EndpointHeader.jsx'
 
 {/* :::note
 See also our [Postman docs](https://api.docs.verified.inc/).
